+++ applications/harnesses/KN_loderunner/train_LodeRunner_ddp.py
@@ -8,27 +8,25 @@ import torch.distributed as dist
 from torch.nn.parallel import DistributedDataParallel as DDP
 from torch.optim.lr_scheduler import LambdaLR
 
-from yoke.models.vit.swin.bomberman import LodeRunner
-from yoke.datasets.lsc_dataset import LSC_rho2rho_temporal_DataSet
-from yoke.utils.training.epoch.loderunner import train_DDP_scalar_temporal_loderunner_epoch_gri
+from yoke.models.vit.swin.bomberman import (
+    LodeRunner,
+    ScalarTemporalConditionedLodeRunner_gri,
+)
+from yoke.datasets.kilonova_dataset import (
+    Kilonova_lc_scalar_context_DataSet_gri,
+    load_or_compute_band_normalization,
+)
+from yoke.utils.training.epoch.loderunner import (
+    train_DDP_scalar_temporal_loderunner_epoch_gri,
+)
 from yoke.utils.restart import continuation_setup
 from yoke.utils.dataload import make_distributed_dataloader
 from yoke.utils.checkpointing import load_model_and_optimizer
 from yoke.utils.checkpointing import save_model_and_optimizer
+from yoke.utils.checkpointing import load_direct_loderunner_checkpoint
+from yoke.utils.parallel import setup_distributed, cleanup_distributed
 from yoke.lr_schedulers import CosineWithWarmupScheduler
 from yoke.helpers import cli
-
-# FIXME remove if restructure
-from torch.utils.data import Dataset, DataLoader, random_split
-import glob
-import random
-
-#MEAN = 24.694652705328807
-#STD = 4.67030961432848
-
-GLOBAL_GMAG_MEAN = 24.694652705328807
-GLOBAL_GMAG_STD  = 4.67030961432848
-EPS = 1e-6
 
 #############################################
 # Inputs
@@ -61,598 +59,6 @@ parser.set_defaults(
     validation_filelist="lsc240420_prefixes_validation_10pct.txt",
     test_filelist="lsc240420_prefixes_test_10pct.txt",
 )
-
-
-def compute_band_normalization(
-    file_prefix_list,
-    band_keys=("arr_ztfg", "arr_ztfr", "arr_ztfi"),
-    value_col=1,
-    stats_path="kilonova_gri_norm_stats.npz",
-):
-    """
-    Compute global per-band mean/std over the training files only.
-
-    Saves:
-        means: shape [3]
-        stds:  shape [3]
-    """
-
-    sums = np.zeros(len(band_keys), dtype=np.float64)
-    sums_sq = np.zeros(len(band_keys), dtype=np.float64)
-    counts = np.zeros(len(band_keys), dtype=np.float64)
-
-    for fn in file_prefix_list:
-        data = np.load(fn, allow_pickle=True)
-
-        for b, key in enumerate(band_keys):
-            vals = data[key][:, value_col].astype(np.float64)
-
-            finite = np.isfinite(vals)
-            vals = vals[finite]
-
-            sums[b] += vals.sum()
-            sums_sq[b] += np.square(vals).sum()
-            counts[b] += vals.size
-
-        data.close()
-
-    means = sums / counts
-    variances = sums_sq / counts - means**2
-    variances = np.maximum(variances, 1e-12)
-    stds = np.sqrt(variances)
-
-    means = means.astype(np.float32)
-    stds = stds.astype(np.float32)
-
-    np.savez(
-        stats_path,
-        means=means,
-        stds=stds,
-        band_keys=np.array(band_keys),
-        value_col=value_col,
-    )
-
-    print("Saved normalization stats:", stats_path)
-    print("means:", means)
-    print("stds:", stds)
-
-    return means, stds
-
-
-def load_or_compute_band_normalization(
-    stats_path="kilonova_gri_norm_stats.npz",
-    band_keys=("arr_ztfg", "arr_ztfr", "arr_ztfi"),
-    value_col=1,
-):
-    file_prefix_list = sorted(
-        glob.glob(
-            "/net/sescratch1/atoivonen/data/KN_lightcurves/uniform_dataset_20000/lc_*.npz"
-        )
-    )
-
-    if os.path.exists(stats_path):
-        stats = np.load(stats_path, allow_pickle=True)
-        means = stats["means"].astype(np.float32)
-        stds = stats["stds"].astype(np.float32)
-        stats.close()
-
-        print("Loaded normalization stats:", stats_path)
-        print("means:", means)
-        print("stds:", stds)
-
-        return means, stds
-
-    return compute_band_normalization(
-        file_prefix_list=file_prefix_list,
-        band_keys=band_keys,
-        value_col=value_col,
-        stats_path=stats_path,
-    )
-
-
-class Kilonova_lc_scalar_context_DataSet_gri(Dataset):
-    def __init__(
-        self,
-        N_imgs=0,
-        context_len=5,
-        band_keys=("arr_ztfg", "arr_ztfr", "arr_ztfi"),
-        value_col=1,
-        means=None,
-        stds=None,
-        predicts_delta=True,
-    ):
-        file_prefix_list = sorted(
-            glob.glob(
-                "/net/sescratch1/atoivonen/data/KN_lightcurves/"
-                "uniform_dataset_20000/lc_*.npz"
-            )
-        )
-
-        if N_imgs == 0:
-            self.file_prefix_list = file_prefix_list
-        else:
-            self.file_prefix_list = list(
-                np.random.choice(file_prefix_list, N_imgs, replace=False)
-            )
-
-        random.shuffle(self.file_prefix_list)
-
-        self.context_len = context_len
-        self.band_keys = tuple(band_keys)
-        self.value_col = value_col
-        self.n_channels = len(self.band_keys)
-        self.predicts_delta = predicts_delta
-
-        if means is None:
-            raise ValueError(
-                "means must be provided for per-band normalization. "
-                "Expected shape [n_channels], e.g. [g_mean, r_mean, i_mean]."
-            )
-
-        if stds is None:
-            raise ValueError(
-                "stds must be provided for per-band normalization. "
-                "Expected shape [n_channels], e.g. [g_std, r_std, i_std]."
-            )
-
-        self.means = np.asarray(means, dtype=np.float32)
-        self.stds = np.asarray(stds, dtype=np.float32)
-
-        if self.means.shape[0] != self.n_channels:
-            raise ValueError(
-                f"means has length {self.means.shape[0]}, "
-                f"but n_channels={self.n_channels}"
-            )
-
-        if self.stds.shape[0] != self.n_channels:
-            raise ValueError(
-                f"stds has length {self.stds.shape[0]}, "
-                f"but n_channels={self.n_channels}"
-            )
-
-        if np.any(self.stds <= 0):
-            raise ValueError(f"All stds must be positive. Got stds={self.stds}")
-
-        self.samples = []
-
-        for file_idx, fn in enumerate(self.file_prefix_list):
-            data = np.load(fn, allow_pickle=True)
-
-            # Assume all bands share the same time grid.
-            mjd = data[self.band_keys[0]][:, 0]
-            n_times = len(mjd)
-
-            data.close()
-
-            max_start = n_times - context_len - 1
-            for startIDX in range(max_start + 1):
-                self.samples.append((file_idx, startIDX))
-
-    def __len__(self):
-        return len(self.samples)
-
-    def __getitem__(self, index):
-        file_idx, startIDX = self.samples[index]
-        fn = self.file_prefix_list[file_idx]
-
-        data = np.load(fn, allow_pickle=True)
-
-        # Time grid from the first band.
-        # Assumes arr_ztfg, arr_ztfr, arr_ztfi have matching MJD columns.
-        mjd = data[self.band_keys[0]][:, 0].astype(np.float32)
-
-        # Stack one scalar value column from each band.
-        # vals shape: [T, 3] for g/r/i.
-        vals = np.stack(
-            [
-                data[key][:, self.value_col].astype(np.float32)
-                for key in self.band_keys
-            ],
-            axis=1,
-        )
-
-        data.close()
-
-        # Per-band normalization.
-        # vals[:, 0] = normalized g
-        # vals[:, 1] = normalized r
-        # vals[:, 2] = normalized i
-        vals = (vals - self.means[None, :]) / (self.stds[None, :] + EPS)
-
-        t0 = mjd.min()
-        t_obs = mjd - t0
-
-        target_idx = startIDX + self.context_len
-
-        # Context values shape: [context_len, 3]
-        context_vals = vals[startIDX:target_idx]
-
-        # Relative observation times shape: [context_len]
-        rel_t = t_obs[startIDX:target_idx] - t_obs[startIDX]
-        rel_t = rel_t.astype(np.float32)
-
-        # Flatten context as:
-        # [g0, r0, i0, g1, r1, i1, ..., gK, rK, iK]
-        context_flat = context_vals.reshape(-1).astype(np.float32)
-
-        # Final input shape:
-        # [(3 * context_len) + context_len]
-        #
-        # For context_len=5:
-        # x.shape == [20]
-        x = np.concatenate([context_flat, rel_t], axis=0)
-        x = torch.tensor(x, dtype=torch.float32)
-
-        if self.predicts_delta:
-            # Predict normalized delta for each band:
-            # [delta_g, delta_r, delta_i]
-            target_vals = vals[target_idx] - vals[target_idx - 1]
-        else:
-            # Predict normalized absolute next value:
-            # [g_next, r_next, i_next]
-            target_vals = vals[target_idx]
-
-        target = torch.tensor(target_vals, dtype=torch.float32)
-
-        Dt = torch.tensor(
-            t_obs[target_idx] - t_obs[target_idx - 1],
-            dtype=torch.float32,
-        )
-
-        return x, target, Dt
-
-
-class Kilonova_lc_scalar_context_DataSet(Dataset):
-    def __init__(
-        self,
-        N_imgs=0,
-        context_len=5,
-    ):
-        file_prefix_list = sorted(
-            glob.glob("/net/sescratch1/atoivonen/data/KN_lightcurves/uniform_dataset_20000/lc_*.npz")
-        )
-
-        if N_imgs == 0:
-            self.file_prefix_list = file_prefix_list
-        else:
-            self.file_prefix_list = list(np.random.choice(file_prefix_list, N_imgs, replace=False))
-
-        random.shuffle(self.file_prefix_list)
-
-        self.context_len = context_len
-        self.samples = []
-
-        for file_idx, fn in enumerate(self.file_prefix_list):
-            data = np.load(fn, allow_pickle=True)
-            mjd = data["arr_ztfg"][:, 0]
-            n_times = len(mjd)
-            data.close()
-
-            max_start = n_times - context_len - 1
-            for startIDX in range(max_start + 1):
-                self.samples.append((file_idx, startIDX))
-
-    def __len__(self):
-        return len(self.samples)
-
-    def __getitem__(self, index):
-        file_idx, startIDX = self.samples[index]
-        fn = self.file_prefix_list[file_idx]
-
-        data = np.load(fn, allow_pickle=True)
-        arr = data["arr_ztfg"]
-
-        mjd = arr[:, 0]
-        g_mag = arr[:, 1].astype(np.float32)
-
-        g_mag = (g_mag - GLOBAL_GMAG_MEAN) / (GLOBAL_GMAG_STD + EPS)
-
-        t0 = mjd.min()
-        t_obs = mjd - t0
-
-        target_idx = startIDX + self.context_len
-
-        mags = g_mag[startIDX:target_idx].astype(np.float32)
-
-        # context_len values; first dt is 0, remaining are relative times
-        rel_t = t_obs[startIDX:target_idx] - t_obs[startIDX]
-        rel_t = rel_t.astype(np.float32)
-
-        x = np.concatenate([mags, rel_t], axis=0)
-        x = torch.tensor(x, dtype=torch.float32)
-
-        delta_mag = g_mag[target_idx] - g_mag[target_idx - 1]
-        target = torch.tensor(delta_mag, dtype=torch.float32)
-
-        Dt = torch.tensor(
-            t_obs[target_idx] - t_obs[target_idx - 1],
-            dtype=torch.float32,
-        )
-
-        # Keep target image-shaped only if your datastep still expects image targets.
-        # Better is to update datastep to use scalar target directly.
-        data.close()
-        return x, target, Dt
-
-
-class ScalarTemporalConditionedLodeRunner_gri(nn.Module):
-    def __init__(
-        self,
-        backbone: nn.Module,
-        context_len: int = 5,
-        n_input_channels: int = 3,
-        n_output_channels: int = 3,
-        image_size=(1120, 400),
-        backbone_channels: int = 8,
-        hidden: int = 64,
-    ):
-        super().__init__()
-
-        self.backbone = backbone
-        self.context_len = context_len
-        self.n_input_channels = n_input_channels
-        self.n_output_channels = n_output_channels
-        self.image_size = image_size
-        self.backbone_channels = backbone_channels
-
-        # Dataset x layout:
-        #   [g0, r0, i0, g1, r1, i1, ..., gK, rK, iK, t0, t1, ..., tK]
-        #
-        # input_dim = context_len * n_input_channels + context_len
-        input_dim = context_len * n_input_channels + context_len
-
-        # Maps scalar temporal context into the 8 pseudo-channels expected by
-        # the pretrained LodeRunner backbone.
-        self.conditioner = nn.Sequential(
-            nn.Linear(input_dim, hidden),
-            nn.GELU(),
-            nn.Linear(hidden, hidden),
-            nn.GELU(),
-            nn.Linear(hidden, backbone_channels),
-        )
-
-        # Maps the 8-channel LodeRunner output back to 3 scalar predictions:
-        #   [delta_g, delta_r, delta_i]
-        self.output_head = nn.Sequential(
-            nn.Linear(backbone_channels, hidden),
-            nn.GELU(),
-            nn.Linear(hidden, n_output_channels),
-        )
-
-    def forward(self, x, in_vars, out_vars, Dt):
-        """
-        x: [B, context_len * n_input_channels + context_len]
-
-        For 3-band, context_len=5:
-            x.shape == [B, 20]
-
-        Returns:
-            pred: [B, 3]
-        """
-        B = x.shape[0]
-        H, W = self.image_size
-
-        channel_vals = self.conditioner(x)  # [B, 8]
-
-        pseudo_img = channel_vals.view(
-            B,
-            self.backbone_channels,
-            1,
-            1,
-        ).expand(
-            B,
-            self.backbone_channels,
-            H,
-            W,
-        )
-
-        backbone_in_vars = torch.arange(self.backbone_channels, device=x.device)
-        backbone_out_vars = torch.arange(self.backbone_channels, device=x.device)
-
-        pred_img = self.backbone(
-            pseudo_img,
-            backbone_in_vars,
-            backbone_out_vars,
-            Dt,
-        )  # [B, 8, H, W]
-
-        # Collapse spatial dimensions to 8 backbone-channel summaries.
-        pred_channel_vals = pred_img.mean(dim=(2, 3))  # [B, 8]
-
-        # Convert 8 backbone channels to 3 output bands.
-        pred = self.output_head(pred_channel_vals)  # [B, 3]
-
-        return pred
-
-
-class ScalarTemporalConditionedLodeRunner(nn.Module):
-    def __init__(
-        self,
-        backbone: nn.Module,
-        context_len: int = 5,
-        image_size=(1120, 400),
-        n_channels: int = 8,
-        hidden: int = 64,
-    ):
-        super().__init__()
-        self.backbone = backbone
-        self.context_len = context_len
-        self.image_size = image_size
-        self.n_channels = n_channels
-
-        # mags + dts
-        self.conditioner = nn.Sequential(
-            nn.Linear(2 * context_len, hidden),
-            nn.GELU(),
-            nn.Linear(hidden, hidden),
-            nn.GELU(),
-            nn.Linear(hidden, n_channels),
-        )
-
-    def forward(self, x, in_vars, out_vars, Dt):
-        """
-        x: [B, 2 * context_len]
-           first context_len entries are magnitudes
-           second context_len entries are temporal deltas
-        """
-        B = x.shape[0]
-        H, W = self.image_size
-
-        channel_vals = self.conditioner(x)  # [B, 8]
-
-        pseudo_img = channel_vals.view(B, self.n_channels, 1, 1).expand(
-            B,
-            self.n_channels,
-            H,
-            W,
-        )
-
-        pred_img = self.backbone(pseudo_img, in_vars, out_vars, Dt)
-
-        # Convert LodeRunner image output back to scalar delta prediction
-        pred_scalar = pred_img.mean(dim=(1, 2, 3))
-
-        return pred_scalar
-
-
-def load_direct_loderunner_checkpoint(
-    checkpoint_path,
-    model_args,
-    optimizer_kwargs,
-    device,
-):
-    checkpoint_data = torch.load(
-        checkpoint_path,
-        map_location=device,
-        weights_only=False,
-    )
-
-    saved_model_args = checkpoint_data.get("model_args", model_args)
-    context_len = checkpoint_data.get("context_len", 5)
-
-    backbone = LodeRunner(**saved_model_args).to(device)
-
-    model = ScalarTemporalConditionedLodeRunner_gri(
-        backbone=backbone,
-        context_len=context_len,
-        n_input_channels=checkpoint_data.get("n_input_channels", 3),
-        n_output_channels=checkpoint_data.get("n_output_channels", 3),
-        image_size=saved_model_args["image_size"],
-        backbone_channels=checkpoint_data.get("backbone_channels", 8),
-        hidden=checkpoint_data.get("hidden", 64),
-    ).to(device)
-
-    state_dict = checkpoint_data["model_state_dict"]
-
-    # Remove DDP prefix if present
-    if any(k.startswith("module.") for k in state_dict.keys()):
-        state_dict = {
-            k.replace("module.", "", 1): v
-            for k, v in state_dict.items()
-        }
-
-    # Detect checkpoint type
-    is_wrapper_checkpoint = any(
-        k.startswith("backbone.") for k in state_dict.keys()
-    )
-
-    # -------------------------------------------------
-    # OLD plain LodeRunner checkpoint
-    # -------------------------------------------------
-    if not is_wrapper_checkpoint:
-
-        missing_keys, unexpected_keys = model.backbone.load_state_dict(
-            state_dict,
-            strict=False,
-        )
-
-        print("Loaded old LodeRunner checkpoint into model.backbone")
-        print("Missing backbone keys:", missing_keys)
-        print("Unexpected backbone keys:", unexpected_keys)
-
-        # This is NOT a true continuation.
-        # Conditioner is newly initialized.
-        starting_epoch = 0
-
-    # -------------------------------------------------
-    # NEW ScalarTemporalConditionedLodeRunner checkpoint
-    # -------------------------------------------------
-    else:
-
-        model.load_state_dict(state_dict, strict=True)
-
-        print("Loaded ScalarTemporalConditionedLodeRunner checkpoint")
-
-        starting_epoch = checkpoint_data.get("epoch", 0)
-
-    noise_scale = checkpoint_data.get("noise_scale", 0.0)
-    model.backbone.noise_scale = noise_scale
-
-    # Freeze pretrained backbone
-    for p in model.backbone.parameters():
-        p.requires_grad = False
-
-    # Train conditioner
-    for p in model.conditioner.parameters():
-        p.requires_grad = True
-
-    #optimizer = torch.optim.AdamW(
-    #    model.conditioner.parameters(),
-    #    **optimizer_kwargs,
-    #)
-    
-    optimizer = torch.optim.AdamW(
-        list(model.conditioner.parameters()) +
-        list(model.output_head.parameters()),
-        **optimizer_kwargs,
-    )
-
-    # Only restore optimizer for TRUE continuation checkpoints
-    if (
-        is_wrapper_checkpoint
-        and "optimizer_state_dict" in checkpoint_data
-    ):
-
-        optimizer.load_state_dict(
-            checkpoint_data["optimizer_state_dict"]
-        )
-
-        for state in optimizer.state.values():
-            for key, value in state.items():
-                if isinstance(value, torch.Tensor):
-                    state[key] = value.to(device)
-
-    return model, optimizer, starting_epoch
-
-
-def setup_distributed():
-    # ----- 1) Basic setup & environment variables -----
-    # Rely on Slurm variables: SLURM_PROCID, SLURM_NTASKS, SLURM_LOCALID, etc.
-    rank = int(os.environ["SLURM_PROCID"])  # global rank
-    world_size = int(os.environ["SLURM_NTASKS"])  # total number of processes
-    local_rank = int(os.environ["SLURM_LOCALID"])  # local rank (GPU index on this node)
-
-    master_addr = os.environ["MASTER_ADDR"]
-    master_port = os.environ["MASTER_PORT"]
-
-    # ----- 2) Set the current GPU device for this process -----
-    torch.cuda.set_device(local_rank)
-    device = torch.device(f"cuda:{local_rank}")
-
-    # ----- 3) Initialize the process group -----
-    dist.init_process_group(
-        backend="nccl",
-        init_method=f"tcp://{master_addr}:{master_port}",
-        world_size=world_size,
-        rank=rank,
-    )
-
-    return rank, world_size, local_rank, device
-
-
-def cleanup_distributed():
-    # ----- 8) Clean up (optional) -----
-    dist.destroy_process_group()
 
 
 def main(args, rank, world_size, local_rank, device):
