+++ src/yoke/models/vit/swin/bomberman.py
@@ -332,6 +332,125 @@ class Lightning_LodeRunner(LightningModule):
             self.log("val_loss", batch_loss, sync_dist=True)
 
 
+class ScalarTemporalConditionedLodeRunner_gri(nn.Module):
+    """Scalar-temporal wrapper around a pretrained LodeRunner backbone.
+
+    Maps a scalar temporal context (flattened per-band values plus relative
+    observation times) into the pseudo-channel image expected by a pretrained
+    LodeRunner backbone, then collapses the backbone's spatial output back to a
+    small number of scalar band predictions. Used for the kilonova light-curve
+    (g/r/i) forecasting task.
+
+    Args:
+        backbone (nn.Module): Pretrained LodeRunner backbone.
+        context_len (int): Number of context timesteps.
+        n_input_channels (int): Number of input bands (e.g. 3 for g/r/i).
+        n_output_channels (int): Number of predicted bands.
+        image_size (tuple): Spatial size (H, W) fed to the backbone.
+        backbone_channels (int): Number of pseudo-channels the backbone expects.
+        hidden (int): Hidden width of the conditioner/output-head MLPs.
+    """
+
+    def __init__(
+        self,
+        backbone: nn.Module,
+        context_len: int = 5,
+        n_input_channels: int = 3,
+        n_output_channels: int = 3,
+        image_size: tuple[int, int] = (1120, 400),
+        backbone_channels: int = 8,
+        hidden: int = 64,
+    ) -> None:
+        """Initialize conditioner and output-head around the backbone."""
+        super().__init__()
+
+        self.backbone = backbone
+        self.context_len = context_len
+        self.n_input_channels = n_input_channels
+        self.n_output_channels = n_output_channels
+        self.image_size = image_size
+        self.backbone_channels = backbone_channels
+
+        # Dataset x layout:
+        #   [g0, r0, i0, g1, r1, i1, ..., gK, rK, iK, t0, t1, ..., tK]
+        #
+        # input_dim = context_len * n_input_channels + context_len
+        input_dim = context_len * n_input_channels + context_len
+
+        # Maps scalar temporal context into the 8 pseudo-channels expected by
+        # the pretrained LodeRunner backbone.
+        self.conditioner = nn.Sequential(
+            nn.Linear(input_dim, hidden),
+            nn.GELU(),
+            nn.Linear(hidden, hidden),
+            nn.GELU(),
+            nn.Linear(hidden, backbone_channels),
+        )
+
+        # Maps the 8-channel LodeRunner output back to 3 scalar predictions:
+        #   [delta_g, delta_r, delta_i]
+        self.output_head = nn.Sequential(
+            nn.Linear(backbone_channels, hidden),
+            nn.GELU(),
+            nn.Linear(hidden, n_output_channels),
+        )
+
+    def forward(
+        self,
+        x: torch.Tensor,
+        in_vars: torch.Tensor,
+        out_vars: torch.Tensor,
+        Dt: torch.Tensor,
+    ) -> torch.Tensor:
+        """Forward pass.
+
+        Args:
+            x (torch.Tensor): Scalar temporal context of shape
+                [B, context_len * n_input_channels + context_len]. For 3-band,
+                context_len=5, x.shape == [B, 20].
+            in_vars (torch.Tensor): Kept for LodeRunner API compatibility.
+            out_vars (torch.Tensor): Kept for LodeRunner API compatibility.
+            Dt (torch.Tensor): Lead-time tensor passed to the backbone.
+
+        Returns:
+            pred (torch.Tensor): Predictions of shape [B, n_output_channels].
+        """
+        B = x.shape[0]
+        H, W = self.image_size
+
+        channel_vals = self.conditioner(x)  # [B, 8]
+
+        pseudo_img = channel_vals.view(
+            B,
+            self.backbone_channels,
+            1,
+            1,
+        ).expand(
+            B,
+            self.backbone_channels,
+            H,
+            W,
+        )
+
+        backbone_in_vars = torch.arange(self.backbone_channels, device=x.device)
+        backbone_out_vars = torch.arange(self.backbone_channels, device=x.device)
+
+        pred_img = self.backbone(
+            pseudo_img,
+            backbone_in_vars,
+            backbone_out_vars,
+            Dt,
+        )  # [B, 8, H, W]
+
+        # Collapse spatial dimensions to 8 backbone-channel summaries.
+        pred_channel_vals = pred_img.mean(dim=(2, 3))  # [B, 8]
+
+        # Convert 8 backbone channels to 3 output bands.
+        pred = self.output_head(pred_channel_vals)  # [B, 3]
+
+        return pred
+
+
 if __name__ == "__main__":
     from yoke.utils.parameters import count_torch_params
